refactor: give the image-header reader, the digest and the browser lookup one home each
Three helpers gain a second caller in the Open Graph work, so each moves to a
module both sides import rather than being copied:

- `intrinsicDimensions` and its PNG/SVG header readers move to
  `shared/content/image-dimensions.ts`, with `Dimensions` renamed `Sized` per
  the naming families now that it is public API. A hand-written second PNG
  header reader is exactly the drift the derivation rule exists to stop.
- `mermaidHash` becomes `contentHash` in `shared/content/content-hash.ts`; the
  digest is not mermaid-specific and the OG manifest needs the same one. The
  algorithm is untouched, so every committed render still resolves.
  `mermaid-hash.ts` is renamed `mermaid-renders.ts`, which is what it holds now
  that the digest has left it.
- `findChromium` moves to `scripts/lib/chromium.ts`. The candidate list is
  environment knowledge, and two copies of it would rot separately.

`ContentDocument` gains `ogImageSize`, resolved with `ogImageUrl` in one helper
so the two cannot disagree, and `constructMetadata` publishes it — a custom
card's dimensions are knowable by reading the file, which is what the avatar
special-case had assumed they were not.
