+++ src/shared/content/index.ts
@@ -17,6 +17,7 @@ export {
   type WithContentDocument,
 } from './documents';
 export type { Frontmatter, WithFrontmatter } from './frontmatter';
+export type { Sized, WithOptionalOgImageSize } from './image-dimensions';
 export {
   type DocumentCard,
   type Heading,
